Migrated to multistage build to reduce final image size.
First stage is only responsible to build binaries which are later used
in the second stage to build final image.
Splitting up this work allows to install build dependencies like
libraries, compilers, etc.. only in the first stage. As a result a lot
of space is saved in the final image.

Python tools like httpie+edgegrid use Python wheel to transport binary
packages to the final image.

Go tools don't require any source deps therefore only binaries are
copied to the final image.

CLI modules are not optimized and copied along with sources to the final
image (this may be improved in future).

The above allowed to drop from 1.89GB to 988MB of the image size.

Terraform provider as well as the core binary are commented out to
deliver same subset of tools like original Dockerfile.

Note: Since build tools are not available in the final image it's not
possible to upgrade or install something new. Essentially, this should
not be a problem since where new versions of CLI/tools are release,
a new docker image should be created.